+++ src/main/java/org/markdownwriterfx/dialogs/ImageDialog.java
@@ -41,6 +41,7 @@ import javafx.stage.Window;
 import org.markdownwriterfx.Messages;
 import org.markdownwriterfx.controls.BrowseFileButton;
 import org.markdownwriterfx.controls.EscapeTextField;
+import org.markdownwriterfx.util.Utils;
 import org.tbee.javafx.scene.layout.fxml.MigPane;
 
 /**
@@ -71,6 +72,8 @@ public class ImageDialog
 		dialogPane.lookupButton(ButtonType.OK).disableProperty().bind(
 				urlField.escapedTextProperty().isEmpty()
 					.or(textField.escapedTextProperty().isEmpty()));
+
+		Utils.fixSpaceAfterDeadKey(dialogPane.getScene());
 
 		image.bind(Bindings.when(titleField.escapedTextProperty().isNotEmpty())
 				.then(Bindings.format("![%s](%s \"%s\")", textField.escapedTextProperty(), urlField.escapedTextProperty(), titleField.escapedTextProperty()))